+++ src/ArticleCard/index.js
@@ -1,4 +1,0 @@
-import ArticleCard, { ARTICLE_CARD_LAYOUTS } from './ArticleCard';
-
-export default ArticleCard;
-export { ArticleCard, ARTICLE_CARD_LAYOUTS };