+++ src/FeatureCard/index.js
@@ -1,0 +1,4 @@
+import FeatureCard, { FEATURE_CARD_LAYOUTS } from './FeatureCard';
+
+export default FeatureCard;
+export { FeatureCard, FEATURE_CARD_LAYOUTS };