+++ src/NavigationBar/NavigationItem.js
@@ -4,10 +4,10 @@ import PropTypes from 'prop-types';
 import { Button } from '../Button';
 
 const NavigationItem = props => {
-  const { name, linkUrl, className, ...rest } = props;
+  const { name, className, ...rest } = props;
 
   return (
-    <Button href={linkUrl} bouncy buttonClassName={className} {...rest}>
+    <Button bouncy buttonClassName={className} {...rest}>
       {name}
     </Button>
   );
@@ -15,13 +15,11 @@ const NavigationItem = props => {
 
 NavigationItem.propTypes = {
   name: PropTypes.string,
-  linkUrl: PropTypes.string,
   className: PropTypes.string,
 };
 
 NavigationItem.defaultProps = {
   name: null,
-  linkUrl: null,
   className: null,
 };